chore(ipv6): all logic into ipv6.rs
